+++ R/tidytree.R
@@ -1,0 +1,1 @@
+nodeid.tbl_tree <- utils::getFromNamespace("nodeid.tbl_tree", "tidytree")