+++ NEWS.md
@@ -1,8 +1,9 @@
 # ggtree 1.17.5
 
++ re-implement `geom_strip` with more robust support of labelling strip (2019-08-27, Tue)
 + support `phylog` defined in ade4 package (2019-08-14, Wed)
   - <https://yulab-smu.github.io/treedata-book/chapter9.html#phylog>
-  
+
 # ggtree 1.17.4
 
 + now `geom_cladelabel` supports `unrooted` and `daylight` layouts (2019-08-14, Wed)