+++ NEWS.md
@@ -1,6 +1,8 @@
 # ggtree 1.17.5
 
-+ re-implement `geom_strip` with more robust support of labelling strip (2019-08-27, Tue)
++ update `geom_motif` according to the change of `gggenes` and allow labeling genes (2019-08-27, Tue)
+  - <https://yulab-smu.github.io/treedata-book/chapter11.html#genome-locus>
++ re-implement `geom_strip` with more robust support of labelling strip 
 + support `phylog` defined in ade4 package (2019-08-14, Wed)
   - <https://yulab-smu.github.io/treedata-book/chapter9.html#phylog>